requirements.txt
flask
gunicorn
pillow
pytesseract
opencv-python
numpy
python-dotenv
